# playerSwagger/src/main/java/com/example/playerswagger/dto/PlayerDto.java
package com.example.playerswagger.dto;

import lombok.Data;

@Data
public class PlayerDto {
    private int id;
    private String firstName;
    private String lastName;
    //awdawdawdawda
}